(PC-12750)[ADAGE]: move paceholder in constants and correct last test
