+++ adage-front/src/app/components/OffersInstantSearch/OffersSearch/Offers/Pagination/index.ts
@@ -1,1 +1,1 @@
-export { default } from './Pagination'
+export { Pagination } from './Pagination'
