+++ adage-front/src/app/components/OffersInstantSearch/OffersSearch/SearchBox/SearchBox.scss
@@ -1,4 +1,4 @@
-@import 'sass-rem';
+@import 'node_modules/sass-rem/index';
 @import 'styles';
 
 .search-wrapper {
