+++ adage-front/src/app/components/OffersInstantSearch/OffersSearch/SearchBox/constants.ts
@@ -1,0 +1,1 @@
+export const placeholder = 'Nom de l’offre ou du partenaire culturel'
