+++ adage-front/src/app/components/OffersInstantSearch/OffersSearch/SearchBox/index.ts
@@ -1,0 +1,1 @@
+export * from './constants'
